ml/requirements.txt
amulet-core==1.9.11
click==8.1.3
cloup==2.0.0.post1
gdpc==6.1.1
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
psutil==5.9.5
pytorch-symbolic==1.0.7
pyglm==2.7.0
scikit-learn==1.2.2
termcolor==2.1.1
torch==2.0.0
tqdm==4.64.1
umap-learn==0.5.3
wandb==0.15.2
